+++ doc/changelog.rst
@@ -1,9 +1,13 @@
-Development 2.9.6
+Development 2.9.7
 =================
 
 Enhancements:
 
 * pff2mt: speedup operation by 70%+
+* emsmdb: strike limits (raise to infinity) for session handles, user handles
+  and notify handles
+* emsmdb: new configuration directives ems_max_active_notifh,
+  ems_max_active_sessions, ems_max_active_users
 
 
 Gromox 2.9 (2023-05-10)